address builds where sizeof != int

+++ server/dhcp.c
@@ -1053,8 +1053,9 @@ void dhcpinform (packet, ms_nulltp)
 		i = d1.len;
 		if (i >= sizeof(raw.file)) {
 			log_info("file name longer than packet field "
-				 "truncated - field: %d name: %d %.*s", 
-				 sizeof(raw.file), i, i, d1.data);
+				 "truncated - field: %lu name: %d %.*s", 
+				 (unsigned long)sizeof(raw.file), i, i,
+				 d1.data);
 			i = sizeof(raw.file);
 		} else
 			raw.file[i] = 0;
@@ -1072,8 +1073,9 @@ void dhcpinform (packet, ms_nulltp)
 		i = d1.len;
 		if (i >= sizeof(raw.sname)) {
 			log_info("server name longer than packet field "
-				 "truncated - field: %d name: %d %.*s", 
-				 sizeof(raw.sname), i, i, d1.data);
+				 "truncated - field: %lu name: %d %.*s", 
+				 (unsigned long)sizeof(raw.sname), i, i,
+				 d1.data);
 			i = sizeof(raw.sname);
 		} else
 			raw.sname[i] = 0;
@@ -2856,9 +2858,10 @@ void dhcp_reply (lease)
 				(sizeof raw.file) - state -> filename.len);
 		else 
 			log_info("file name longer than packet field "
-				 "truncated - field: %d name: %d %.*s", 
-				 sizeof(raw.file), state->filename.len,
-				 state->filename.len, state->filename.data);
+				 "truncated - field: %lu name: %d %.*s", 
+				 (unsigned long)sizeof(raw.file),
+				 state->filename.len, state->filename.len,
+				 state->filename.data);
 	} else
 		bufs |= 1;
 
@@ -2874,8 +2877,9 @@ void dhcp_reply (lease)
 				(sizeof raw.sname) - state -> server_name.len);
 		else 
 			log_info("server name longer than packet field "
-				 "truncated - field: %d name: %d %.*s", 
-				 sizeof(raw.sname), state->server_name.len,
+				 "truncated - field: %lu name: %d %.*s", 
+				 (unsigned long)sizeof(raw.sname),
+				 state->server_name.len,
 				 state->server_name.len,
 				 state->server_name.data);
 	} else
